requirements.txt
absl-py
accelerate
addict==2.4.0
aiohappyeyeballs
aiohttp
aiohttp-cors==0.8.1
aiosignal
annotated-types==0.7.0
antlr4-python3-runtime
asciitree==0.3.3
asttokens
async-timeout
attrs
av
beautifulsoup4==4.13.4
blessed==1.21.0
blinker
boto3
botocore
Brotli
cached-property
cachetools
certifi
cffi
charset-normalizer
click==8.1.8
cloudpickle
colorama
colorful==0.5.7
comm==0.2.2
ConfigArgParse==1.7.1
contourpy
cryptography
cycler
Cython
cytoolz
dash==3.1.1
dask
dataclasses
datasets
debugpy
decorator
diffusers
diffusion_policy==0.0.0
dill
distlib==0.3.9
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
egl_probe==1.0.2
einops
entrypoints
etils==1.5.2
evdev==1.9.2
exceptiongroup
executing
fasteners==0.15
fastjsonschema==2.21.1
filelock
Flask==3.1.1
fonttools
free-mujoco-py==2.1.6
frozenlist
fsspec
fvcore
gdown==5.2.0
gitdb
GitPython
glfW==1.12.0
google-api-core==2.25.1
google-auth
google-auth-oauthlib
googleapis-common-protos==1.63.1
gpustat==1.1.1
grpcio
gym
h5py
huggingface_hub
hydra-core
idna
imagecodecs==2022.9.26
imageio
imageio-ffmpeg
importlib_metadata
importlib_resources
iopath
ipykernel
ipython
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi
Jinja2==3.1.6
jmespath
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client
jupyter_core
jupyterlab_widgets==3.0.15
kiwisolver
labmaze==1.0.6
llvmlite
locket
lxml==6.0.0
Markdown
MarkupSafe
matplotlib
matplotlib-inline
monotonic==1.6
msgpack
mujoco==3.3.3
multidict
multiprocess
munkres==1.1.4
narwhals==1.45.0
nbformat==5.10.4
nest_asyncio
networkx
numba
numcodecs
numpy
nvidia-ml-py==12.575.51
oauthlib
omegaconf
open3d==0.19.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.6.0
packaging
pandas
parameterized==0.9.0
parso
partd
pathtools==0.1.2
pexpect
pickleshare
pillow==11.3.0
piper_sdk==0.3.0
platformdirs
plotly==6.2.0
ply
portalocker
prometheus-client==0.13.1
promise
prompt_toolkit
propcache
proto-plus==1.26.1
protobuf==3.19.6
psutil
ptyprocess
pure_eval
py-spy==0.4.0
pyarrow
pyasn1
pyasn1_modules
pybind11
pybind11-stubgen==2.5.4
pybind11_global
pybullet-svl==3.1.6.4
pycparser
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.1.2
Pygments
PyJWT
pymunk
pynput==1.8.1
PyOpenGL==3.1.9
pyOpenSSL
pyparsing
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyquaternion==0.9.9
PySocks
python-can==4.5.0
python-dateutil
python-xlib==0.33
pytorch3d==0.7.0
pytorchvideo==0.1.5
pytz
pyu2f
PyWavelets==1.6.0
PyYAML
pyzmq
r3m @ https://github.com/facebookresearch/r3m/archive/b2334e726887fa0206962d7984c69c5fb09cceab.tar.gz
ray==2.2.0
referencing==0.36.2
regex
requests
requests-oauthlib
responses
retrying==1.4.0
robomimic==0.2.0
robosuite @ https://github.com/cheng-chi/robosuite/archive/277ab9588ad7a4f4b55cf75508b44aa67ec171f0.tar.gz
rpds-py==0.26.0
rsa
s3transfer
scikit-image
scikit-learn==1.6.1
scikit-video
scipy
sentry-sdk
setproctitle
Shapely
shortuuid
sip
six
smart_open==7.3.0
smmap
soupsieve==2.7
stack_data
tabulate
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
termcolor
threadpoolctl
tifffile
toml
tomli
toolz
torch
torchvision
tornado
tqdm
traitlets
typing-inspection==0.4.1
typing_extensions
tzdata
unicodedata2
urdf-parser-py
urllib3
virtualenv==20.31.2
wandb
wcwidth
Werkzeug
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash
yacs
yarl
zarr
zipp